backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.7
distro==1.9.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mangum==0.19.0
openai==1.57.4
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
setuptools==75.1.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.44.0
langchain == 0.3.12
langchain-core == 0.3.25
langchain-openai == 0.2.12
langchain-text-splitters == 0.3.3
langchain-upstage == 0.4.0
langchain_pinecone == 0.2.0
pinecone == 5.3.1
langgraph == 0.2.59
google-search-results
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
google_search_results==2.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-huggingface==0.1.2
langchain-openai==0.2.12
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.3
langchain-upstage==0.4.0
langgraph==0.2.59
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.7
mangum==0.19.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.57.4
orjson==3.10.13
packaging==24.2
pillow==11.0.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
pydantic==2.10.3
pydantic-settings==2.7.1
pydantic_core==2.27.1
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.44.0
yarl==1.18.3
